+++ src/controllers/Product.js
@@ -1,4 +1,5 @@
 import mongoose from 'mongoose';
+import validations from '../validations/validations';
 
 const ProductModel = mongoose.model('Product');
 
@@ -18,6 +19,10 @@ class Product {
     try {
       const { search } = req.query;
 
+      if (validations.validationField(search)) {
+        return res.status(400).json({ message: 'Informações inválidas' });
+      }
+
       const products = await ProductModel.find({
         title: {
           $regex: search,
@@ -35,6 +40,10 @@ class Product {
   async findProductsByCategory(req, res) {
     try {
       const { search } = req.query;
+
+      if (validations.validationField(search)) {
+        return res.status(400).json({ message: 'Informações inválidas' });
+      }
 
       const products = await ProductModel.find({
         category: {
@@ -55,6 +64,14 @@ class Product {
         title, description, price, category,
       } = req.body;
 
+      if (
+        validations.validationField(title)
+        || validations.validationField(description)
+        || validations.validationField(price)
+      ) {
+        return res.status(400).json({ message: 'Informações inválidas' });
+      }
+
       const newProduct = await new ProductModel({
         title, description, price, category,
       });
@@ -74,6 +91,10 @@ class Product {
         id, title, description, price,
       } = req.body;
 
+      if (validations.validationField(id)) {
+        return res.status(400).json({ message: 'Informações inválidas' });
+      }
+
       await ProductModel.findByIdAndUpdate(id, {
         $set: {
           title,
@@ -92,6 +113,10 @@ class Product {
   async deleteProduct(req, res) {
     try {
       const { id } = req.params;
+
+      if (validations.validationField(id)) {
+        return res.status(400).json({ message: 'Informações inválidas' });
+      }
 
       await ProductModel.findByIdAndDelete({ _id: id });
       return res.status(200).json({ message: 'Produto deletado com sucesso.' });
